Configure SWC compiler to import react runtime by default (#341)
### What's here?

This allows our newer storybook samples to run without having to import
React in the file heading


![image](https://github.com/user-attachments/assets/5bb2f74e-e438-4bad-a2fd-0f0ffafc6827)

This also removes the need to import react from older examples

### Weird caveats

NGS Airline Selector does not seem to want to let go of the import even
after the update. So I kept it there. Pretty bizaare!


![image](https://github.com/user-attachments/assets/85053264-5cd7-4420-a852-ea7c16e26daa)

This is not the case for all the other components

### References

The revisions applied are taken mostly from the storybook documentation
here:


https://storybook.js.org/docs/configure/integration/compilers#the-swc-compiler-doesnt-work-with-react